% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.age.column.R
\name{add.age.col}
\alias{add.age.col}
\title{add age column}
\usage{
add.age.col(data, stage = "Mean")
}
\description{
@param data Input data
       stage At what stage should the ages be calculated? "X","BG","Mean","corr", tbc
}
\keyword{outlier,}
\keyword{range}
\keyword{sigma,}
