% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.age.error.column.R
\name{add.age.sd.col}
\alias{add.age.sd.col}
\title{add age uncertainty column}
\usage{
add.age.sd.col(data, n.sample = 1000)
}
\arguments{
\item{data}{Input data}

\item{n.sample}{Size of sampling for random numbers}
}
\description{
add age uncertainty column
}
\keyword{outlier,}
\keyword{range}
\keyword{sigma,}
