% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.ratios.R
\name{add.ratios}
\alias{add.ratios}
\title{Add Ratios}
\usage{
add.ratios(data, stage = "BG", ox = FALSE)
}
\arguments{
\item{data}{Data frame with original data.
stage The ratios are supposed to be build on what stage? "X","BG","Mean",tbc
ox Boolean to say whether this is an oxid-measurement.}
}
\description{
This function adds measured and activity ratios to your data frame.
}
\examples{
add.ratios(data)
}
\keyword{ratios}
