% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.error.R
\name{age.error}
\alias{age.error}
\title{Age Error}
\usage{
age.error(data)
}
\arguments{
\item{data}{Data with errors to calculate the age errors.}
}
\description{
This function calculates the age errors by using a MC method.
}
\examples{
age.error(data)
}
\keyword{age,}
\keyword{error}

