% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.measurements.R
\name{average.measurements}
\alias{average.measurements}
\title{Average the measurements}
\usage{
average.measurements(data)
}
\arguments{
\item{data}{Data input}
}
\description{
This function combines the data of each individual measurement and adds the standard deviation as an additional column.
}
\examples{
average.measurements(data)
}
\keyword{average}
\keyword{mean,}

