% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{average}
\alias{average}
\title{Average measurements}
\usage{
average(d, clear.outliers = TRUE)
}
\arguments{
\item{data}{Data input}
}
\description{
This function averages the measurements and returns a list of average values and standard deviations.
}
\examples{
average(data)
}
\keyword{average,}
\keyword{mean,}
\keyword{sd}
