% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundCorrection.R
\name{background.correction}
\alias{background.correction}
\title{Function to correct the background}
\usage{
background.correction(d, is.bg, exp = FALSE)
}
\arguments{
\item{d}{Vector of data.}

\item{is.bg}{Vector of same length as d with boolean entries.}

\item{exp}{boolean weather to use an exponential model istead of a linear model. Default is FALSE.}
}
\description{
This function subtracts the background based on a linear or exponential fit.
}
\examples{
background.correction(d = data, is.bg = bg)
}
\keyword{background,}
\keyword{bg,}
\keyword{correction,}
\keyword{exponential}
\keyword{fit}
\keyword{fit,}
\keyword{linear}
