% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.ratios.R
\name{calc.ratios}
\alias{calc.ratios}
\title{Calculate ratios}
\usage{
calc.ratios(d, denominator)
}
\arguments{
\item{d}{Vector of data.}

\item{info}{data.frame with the columns ID with autamic column names and 
Rel.abund with the relative abundance of the isotope}
}
\description{
This function calculates ratios based on the column, that it receives as input.
}
\examples{
calc.conc(d = data, info = isotopeInfo)
}
\keyword{abundance}
\keyword{concentration,}
\keyword{relative}
