% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeDenominator.R
\name{changeDenom}
\alias{changeDenom}
\title{Change denominator}
\usage{
changeDenom(d, new.denominator)
}
\arguments{
\item{d}{Vector of data.}

\item{new.denominator}{string of the new denominator name (eg: "23Na" or "X23Na")}
}
\description{
This function changes ratios based on the column, that it receives as input.
}
\examples{
changeDenom(d = data, new.denominator = "23Na")
}
\keyword{abundance,}
\keyword{change,}
\keyword{concentration,}
\keyword{denominator}
\keyword{relative}
