% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.edges.R
\name{clean.edges}
\alias{clean.edges}
\title{Clean Edges}
\usage{
clean.edges(data)
}
\arguments{
\item{data}{Input data}
}
\description{
This function deletes the first and last entry of each measurement.
}
\examples{
clean.edges(data)
}
\keyword{clean}
\keyword{edges}
\keyword{memory,}
\keyword{uptake,}

