% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.outliers.R
\name{clear.outliers}
\alias{clear.outliers}
\title{Clear Outliers}
\usage{
clear.outliers(x, a = 2)
}
\arguments{
\item{x}{Input data
a Sigma range}
}
\description{
This function deletes the entries which deviate more than a specific sigma range from the mean.
}
\examples{
clear.outliers(x = data, a = 2)
}
\keyword{outlier,}
\keyword{sigma}

