% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2standard.R
\name{compare2standard}
\alias{compare2standard}
\title{compare to standard}
\usage{
compare2standard(d, stdData, denColumn, sampleList)
}
\arguments{
\item{d}{Data input}

\item{stdData}{Data of the composition of the standard materials}

\item{denColumn}{the name of the column, that is used as the denominator}

\item{sampleList}{the list of the standards, that were used}
}
\description{
This function calculates the elemental fractionation.
}
\examples{
average(data)
}
\keyword{average,}
\keyword{mean,}
\keyword{sd}
