% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_exon_3end_couplings.R
\name{compute_exon_3end_couplings}
\alias{compute_exon_3end_couplings}
\title{Compute couplings between 3'ends and exons}
\usage{
compute_exon_3end_couplings(junctions, junctionRef)
}
\arguments{
\item{junctions:}{junctions obtained from Long read data assigned to read_ids}

\item{junctionRef:}{set of junctions of interest classified using SaiLoR}
}
\value{
couplings of 3'ends to exons
}
\description{
Compute couplings between 3'ends and exons
}
