% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_exon_tss_couplings.R
\name{compute_exon_tss_couplings}
\alias{compute_exon_tss_couplings}
\title{Compute couplings exon junctions and Transcription start sites}
\usage{
compute_exon_tss_couplings(junctions, junctionRef)
}
\arguments{
\item{junctions:}{junction data with to junction-reads assignments}

\item{read_ids:}{read_ids for full length reads}

\item{pathJunctionRef:}{reference database used.}
}
\value{

}
\description{
Compute couplings exon junctions and Transcription start sites
}
