% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2counts.R
\name{convert2counts}
\alias{convert2counts}
\title{Convert to Counts}
\usage{
convert2counts(data, cups = get.cups())
}
\arguments{
\item{data}{Input data
cups On which cup the Ions were measured.  Default is no entry needed.}
}
\description{
This function converts all voltages to counts.
}
\examples{
convert2counts(data)
}
\keyword{counts}
\keyword{volts,}
