% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.drift.R
\name{correct.drift}
\alias{correct.drift}
\title{correct drift}
\usage{
correct.drift(data, standard.measurements = c(1, 3), stage = "BG")
}
\arguments{
\item{data}{Data frame with original data.
standard.measurements A list of numbers which were identified to be standards.
stage The ratios are supposed to be build on what stage? "X","BG","Mean","corr", tbc}
}
\description{
This function drift, by taking into account how a standard was repeatedly measured.
}
\examples{
add.ratios(data)
}
\keyword{ratios}
