% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sailor_full_lengths.R
\name{countLinks}
\alias{countLinks}
\title{Count full length reads}
\usage{
countLinks(alignments, linksDatabase)
}
\arguments{
\item{alignments:}{alignments file from minimap2}

\item{linksDatabase:}{created with prepare_links_database}
}
\value{

}
\description{
Count full length reads
}
