% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.graphics.path.R
\name{create.graphics.path}
\alias{create.graphics.path}
\title{Create a path to the desired file where all graphics are stored}
\usage{
create.graphics.path(filename = "RPlot", replace = F)
}
\arguments{
\item{filename}{A string which is going to be the name the output file. Default value is "RPlot"
replace Boolean to indicate whether to check if the file should be replaced if it already exists. Default value is FALSE.}
}
\description{
This function creates a filepath to save a graphic to the folder where all graphics are stored for my masters thesis.
The file can be replaced if the parameter 'replace' is set to TRUE. Otherwise this function checks whethter this file already exists
and alters the filename if necessary.
}
\examples{
create.graphics.path(filename = "Signal_intensities", replace = F)

This creates a new file called Signal_intensities.tex in the folder where the graphics for the Tex-file are stored.
If there is already a file with this name it is changed to Signal_intensities_1.tex.
}
\keyword{file}
\keyword{graphics,}
\keyword{latex}
\keyword{path,}
