% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.folder.R
\name{create.new.measurement.folder}
\alias{create.new.measurement.folder}
\title{Create a new folder for a measurement}
\usage{
create.new.measurement.folder()
}
\description{
This function creates a new folder with the subfolders "data" and "plots".
}
\examples{
create.new.measurement()
}
\keyword{folder,}
\keyword{measurement}
