% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference_junctions.R
\name{create_reference_junctions}
\alias{create_reference_junctions}
\title{Title}
\usage{
create_reference_junctions(pathSJ.out, min.jcounts, refAnnot, type)
}
\value{

}
\description{
Title
}
