% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.measurements.R
\name{find.measurements}
\alias{find.measurements}
\title{Find Measurements}
\usage{
find.measurements(data, threshold = 10)
}
\arguments{
\item{data}{Data input.
threshold Maximum deviation from the natural ratio 238U/235U.}
}
\description{
This function appoints numbers to each measurement.
}
\examples{
find.measurement(data, threshold=5)
}
\keyword{find}
\keyword{measurements,}

