% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_analysis.R
\name{first.analysis}
\alias{first.analysis}
\title{Read raw data from the icap}
\usage{
first.analysis(
  matrix.element = "Ca",
  output.filename = "results",
  subtract.bg = T,
  dismiss.beginning = 10,
  dismiss.after = 10,
  step = 500,
  exp.bg = F
)
}
\arguments{
\item{output.filename}{string default filename of output}

\item{subtract.bg}{boolean whether backround is already subtracted, default = T}

\item{dismiss.beginning}{integer no of measurement points discarded}

\item{dismiss.after}{integer no of measurement points discarded}

\item{step}{integer no of measurement points displayed}

\item{exp.bg}{boolean approximate background with an axponential curve, default: F}
}
\description{
Read raw data from the icap
}
\keyword{data}
\keyword{rawdata,}
\keyword{read}
