% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floatingMedian.R
\name{floatingMedian}
\alias{floatingMedian}
\title{my own floating median}
\usage{
floatingMedian(data, l = 6, clear.outliers = FALSE, pad = FALSE)
}
\arguments{
\item{data}{vector of data}

\item{l}{number of points to be averaged}

\item{clear.outliers}{Boolean}

\item{pad}{boolean whether NA should be added in front and in the back to maintain the same length}
}
\description{
This function gives the floating median with a given number of data points
}
\keyword{floating}
\keyword{sd}
