% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floatingSD.R
\name{floatingSD}
\alias{floatingSD}
\title{my own floating SD}
\usage{
floatingSD(data, l = 6, clear.outliers = FALSE, pad = FALSE)
}
\arguments{
\item{data}{vector of data}

\item{l}{number of points to be averaged}

\item{clear.outliers}{Boolean}

\item{pad}{boolean whether NA should be added in front and in the back to maintain the same length}
}
\description{
This function gives the SD to a floating mean
}
\keyword{floating}
\keyword{sd}
