% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference_junctions.R
\name{get_junctions_from_annot}
\alias{get_junctions_from_annot}
\title{Annotate junctions}
\usage{
get_junctions_from_annot(AnnotJunctions, linksDatabase, genesMultipleJunctions)
}
\arguments{
\item{AnnotJunctionsL:}{junction object from make_annotation_txf}

\item{linksDatabase:}{database to identify TSS and TES to filter exons}

\item{genesMultipleJunctions:}{object of genes with multiple isoforms}
}
\value{

}
\description{
Annotate junctions
}
