% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden.age.search.R
\name{golden.age.search}
\alias{golden.age.search}
\title{Golden Age Search}
\usage{
golden.age.search(data, young.limit = 10, old.limit = 5e+05,
  tolerance = 1, initial30.32 = 5)
}
\arguments{
\item{data}{Data frame as input
young.limit Lower boundary of age in years. Default is 10.
old.limit Upper boundary of age in years. Default is 500,000.
tolerance Stop iteration, if boundaries are less tolerance from each other. Default is 1.
initial30.32 Value for initial 230Th/232Th value. Default is 5.}
}
\description{
This function calculates the age from the input data.
}
\examples{
golden.age.search(data,tolerance=10)
}
\keyword{age,}
\keyword{decay}
\keyword{equation}
\keyword{golden}
\keyword{search,}

