% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.meas.and.bg.R
\name{id.meas.and.bg}
\alias{id.meas.and.bg}
\title{Identify measurement and background}
\usage{
id.meas.and.bg(d, dismiss.beginning = 0, dismiss.after = 0, step = 100)
}
\arguments{
\item{d}{Vector of data.}
}
\description{
This function takes a vector of data, displays it and lets you click on the start and end points of a measurement.
It returns a data frame with a column indicating measurement values and one indicating background values
}
\examples{
read.exp()
}
\keyword{read}
