% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laser.experiment.R
\name{import.LaserExperiment}
\alias{import.LaserExperiment}
\title{Import information from Laser Experiment}
\usage{
import.LaserExperiment(file = NA)
}
\description{
This function reads the csv file of a laser experiment
}
\examples{
import.LaserExperiment()
}
\keyword{info}
\keyword{laser}
