% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracycle.correction.R
\name{intracycle.correction}
\alias{intracycle.correction}
\title{Intracycle correction}
\usage{
intracycle.correction(data)
}
\arguments{
\item{data}{Data input}
}
\description{
This function corrects differences between the cycles by correcting for 238U values.
}
\examples{
intracycle.correction(data)
}
\keyword{intracycle}

