% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOZ.correction.R
\name{KOZ.correct}
\alias{KOZ.correct}
\title{KOZ correction}
\usage{
KOZ.correct(data, standard.measurements = c(1, 3), stage = "mbc")
}
\arguments{
\item{data}{Data frame with original data.
standard.measurements A list of numbers which were identified to be standards.
stage The ratios are supposed to be build on what stage? "X","BG","Mean","corr", tbc}
}
\description{
This function corrects the actvity ratios by taking the standard measurement of KOZ into account
}
\examples{
KOZ.correct(data)
}
\keyword{ratios}
