\name{laser-internals}
\alias{laser-internals}
\alias{yuleint2}
\alias{yuleint2New}
\alias{combinations}
\alias{Ibd}
\alias{IDDX}
\alias{IDDL}
\alias{IpureBirth}
\alias{Iyule2rate}
\alias{Iyule3rate}
\alias{checkbasal}
\alias{gsr}
\alias{node.sons}
\alias{getLikelihood.EXVAR}
\alias{getLikelihood.SPVAR}
\alias{getLikelihood.BOTHVAR}
\alias{pTt.BOTHVAR}
\alias{pTt.EXVAR}
\alias{pTt.SPVAR}
\alias{rhoFxn.BOTHVAR}
\alias{rhoFxn.SPVAR}
\alias{rhoFxn.EXVAR}
\alias{muFx}
\alias{lambdaFx}
\alias{getSpeciationtimes}



\title{ internal functions for package laser }
\description{
  internal functions for \code{laser}
}

\details{
  These functions are for internal use and should not be called by the user
}


\author{ Dan Rabosky \email{drabosky@umich.edu} }

\keyword{ internal }
