\name{laser-package}
\alias{laser-package}
\docType{package}
\title{
Likelihood Analysis of Speciation/Extinction Rates from Phylogenies
}
\description{
\code{laser} implements maximum likelihood methods based on the birth-death
process to test whether diversification rates have changed over time.  The package 
permits batch processing of phylogenies to generate null distributions of test statistics
and posterior distributions of parameter estimates.  Additional functions for manipulating 
branching times from molecular phylogenies and for simulating branching times are provided.
}

\author{
Dan Rabosky \email{DLR32@cornell.edu}
}
\references{
Rabosky, D. L. 2006.  Likelihood methods for inferring temporal shifts in diversification rates.
}
\keyword{ package }


