% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.csv.R
\name{load.csv}
\alias{load.csv}
\title{Function to load csv files}
\usage{
load.csv(filename = NA)
}
\arguments{
\item{filename}{Filename of the csv file to load without ".csv". If not specified you are asked to choose an option, when there is more than one file}
}
\description{
This function allows you to load data from csv-files and put them into a data.frame.
}
\examples{
load.csv()
}
\keyword{csv}
\keyword{read,}
