% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaded_data_2_list.R
\name{loaded_data_2_list}
\alias{loaded_data_2_list}
\title{Average measurements}
\usage{
loaded_data_2_list(data)
}
\arguments{
\item{data}{Data input}
}
\description{
This function averages the measurements and returns a list of average values and standard deviations. The
}
\examples{
loaded_data_2_list(data = read.csv("results.csv"))
}
\keyword{average,}
\keyword{mean,}
\keyword{sd}
