% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_to_junctions.R
\name{make_junction_database}
\alias{make_junction_database}
\title{merge junctions to reads}
\usage{
make_junction_database(bamJunctions)
}
\arguments{
\item{bamJunctions}{}
}
\value{

}
\description{
merge junctions to reads
}
