% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass4column.R
\name{mass4column}
\alias{mass4column}
\title{Clear Outliers}
\usage{
mass4column(data)
}
\arguments{
\item{data}{Data input}
}
\description{
This function appoints masses to each column and returns a seperate vector.
}
\examples{
mass4column(data)
}
\keyword{mass}
