% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb58.correction.R
\name{mb58.correction}
\alias{mb58.correction}
\title{Mass Bias Correction}
\usage{
mb58.correction(data, stage = "corr")
}
\arguments{
\item{data}{Data input.}
}
\description{
This function corrects for mass bias based on an exponential law and the natural ratio of 238U/235U.
}
\examples{
mb58.correction(data)
}
\keyword{bias}
\keyword{mass}
