% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxhy.R
\name{oxhy}
\alias{oxhy}
\title{Function evaluate oxide/Hydride measurements}
\usage{
oxhy()
}
\description{
This function calculates oxide and hydride ratios after substracting the background.
}
\examples{
oxhy()
}
\keyword{evaluate}
\keyword{oxhy,}
