% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFuerMA.R
\name{plot.for.ma}
\alias{plot.for.ma}
\title{Opens the tikzDevice in the graphics folder of the MA}
\usage{
\method{plot}{for.ma}(filename = "RPlot", width = 4, height = 3, ...)
}
\arguments{
\item{filename}{A string which is going to be the name of the output file. Default value is "RPlot"
@param     width numeric Width of the plot in inches. Default is 4.
@param     height numeric height the plot in inches. Default is 3.}
}
\description{
This function opens and loads the tikzDevice to create a tikz output for a plot. It uses therein the function create.graphics.path()
also from the LaseR-package.
}
\examples{
plot.for.ma(filename = "Signalstärke", width = 4, height = 4)

This creates a new file called Signal_intensities.tex in the folder where the graphics for the Tex-file are stored.
Afterwards the plotting has to be done and in the end the device has to be closed by dev.off().
}
\keyword{device,}
\keyword{file}
\keyword{graphics,}
\keyword{latex}
