% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotexperiment.R
\name{plot.laserexperiment}
\alias{plot.laserexperiment}
\title{Plot information from Laser Experiment}
\usage{
\method{plot}{laserexperiment}(data.table = NA, highlight = 0)
}
\arguments{
\item{data.table}{data frame of laser info}

\item{highlight}{id of track to be highlighted}
}
\description{
This function plots a laser experiment
}
\examples{
plot.laserexperiment()
}
\keyword{info,}
\keyword{laser}
\keyword{plot}
