\name{plotNodeNumbers.phylo.Rd}
\alias{plotNodeNumbers.phylo}

\title{ Plot phylogenetic tree with node numbers  }
\description{
  Wrapper for plot.phylo; plots class 'phylo' trees and displays internal and terminal node numbers
}    
\usage{
plotNodeNumbers.phylo(phy)

}


\arguments{
  \item{phy}{ a class 'phylo' phylogenetic tree}

}
\details{


}
\value{

  
}
\references{ 



}
\author{ Dan Rabosky \email{DLR32@cornell.edu} }


}
\keyword{ models }
