% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sailor_full_lengths.R
\name{prepareForCountStarts}
\alias{prepareForCountStarts}
\title{prepare counts trimming the reads to their most 5' into a 1 nt window for counting}
\usage{
prepareForCountStarts(x, window)
}
\arguments{
\item{x}{alignments: alignments file from minimap2}

\item{window}{window for trimming the reads to their most 5'/3'}
}
\value{

}
\description{
prepare counts trimming the reads to their most 5' into a 1 nt window for counting
}
