% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickie.R
\name{quick.auswertung}
\alias{quick.auswertung}
\title{quick evaluation without drift correction or a standard normalization}
\usage{
quick.auswertung()
}
\description{
You have to have the csv file data.csv
}
\keyword{quick,}
