\name{RandomTrees}
\alias{rtrees50}
\alias{rtrees5}
\docType{data}
\title{ Random Phylogenetic Trees}
\description{
  \samp{rtrees5} and \samp{rtrees50} are sets of 5 and 50 phylogenetic trees of 25 taxa
  generated under a stochastic pure-birth (Yule) process using the program \emph{Phyl-O-Gen}
}
\usage{data(rtrees50)}
\details{
  These data can be used to explore \code{\link{fitdAICrc.batch}} and 
  \code{\link{getBtimes.batch}}
}
\source{
  Data generated using \emph{Phyl-O-Gen} (\url{http://evolve.zoo.ox.ac.uk/software/PhyloGen/main.html})
}

\examples{
data(rtrees50)

}
\keyword{datasets}
