% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.exp.R
\name{read.exp}
\alias{read.exp}
\title{Function to read exp-files}
\usage{
read.exp()
}
\description{
This function allows you to load data from exp-files and put them into a data.frame.
}
\examples{
read.exp()
}
\keyword{read}

