% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.iCAP.R
\name{read.iCAP}
\alias{read.iCAP}
\title{Read iCAP data files}
\usage{
read.iCAP(pattern = ".csv", files = NA)
}
\arguments{
\item{pattern}{The pattern in the file names that connects all measurements you want to read. Default ".csv"}

\item{files}{The files that you want to read. Default NA}
}
\description{
This function loads all iCAP data files and returns one data frame, including this data.
}
\examples{
read.iCAP(pattern = "date1_measurementB")
}
\keyword{csv-files}
\keyword{data,}
\keyword{icap,}
\keyword{read,}
