% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_to_junctions.R
\name{read_refjunction_correction}
\alias{read_refjunction_correction}
\title{Title}
\usage{
read_refjunction_correction(
  bamJunctions,
  shortReadJunctions,
  dist.5ss,
  dist.3ss
)
}
\value{

}
\description{
Title
}
