% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sailor_full_lengths.R
\name{readTSSassignment}
\alias{readTSSassignment}
\title{Title}
\usage{
readTSSassignment(startsAlignemnts, TSSCoordinate.base)
}
\arguments{
\item{startsAlignemnts}{trimmed reads  output from prepareForCountStarts}

\item{TSSCoordinate.base}{reference database for TSS}
}
\value{

}
\description{
Title
}
