% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.identify.save.R
\name{ris}
\alias{ris}
\title{ris (read identify save)}
\usage{
ris(len = 100)
}
\arguments{
\item{len}{Length of the plot in x-direction}
}
\description{
This function reads in the exp file, plots the 238U-Signal and lets you identify where the measurements are and what is background. 
All this is then saved in a csv file.
}
\examples{

}
\keyword{background,}
\keyword{csv,}
\keyword{file}
\keyword{file,}
\keyword{measurements,}
