% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2dates.R
\name{s2date}
\alias{s2date}
\title{Convert the Time stamps to real dates}
\usage{
s2date(df)
}
\arguments{
\item{data}{Data frame}
}
\description{
returns data frame with an additional column "dates" including the dates
}
\examples{
s2date(df = data)
}
\keyword{date,}
\keyword{seconds}
