% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.data.R
\name{save.data}
\alias{save.data}
\title{Function to write data in a csv file}
\usage{
save.data(data, filename = NA)
}
\arguments{
\item{filename}{Filename of the csv file to write ".csv". If not specified you are asked to enter a filename.}
}
\description{
This function allows you to load data from csv-files and put them into a data.frame.
}
\examples{
save.data()
}
\keyword{csv}
\keyword{write,}
