% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{set.parameters}
\alias{set.parameters}
\title{Get cups}
\usage{
set.parameters()
}
\description{
This adds important parameters to the global environment
}
\examples{
set.parameters()
}
\keyword{constants,}
\keyword{decay}
\keyword{masses}
\keyword{parameters,}
