\name{skinkdiversity}
\alias{skinkdiversity}
\docType{data}
\title{ Species richness for Australian sphenomorphine skink genera}
\description{
	Species richness data for genera in \code{skinktree}; also data used for analysis in Rabosky et al. 2007
  }
\usage{data(skinkdiversity)}

\source{
Rabosky, D. L., S. C. Donnellan, A. L. Talaba, and I. J. Lovette. 2007. Exceptional among-lineage variation in diversification rates during the radiation of Australia's largest vertebrate clade. Proc. Roy. Soc. Lond. Ser. B 274:2915-2923.
}
\references{
Rabosky, D. L., S. C. Donnellan, A. L. Talaba, and I. J. Lovette. 2007. Exceptional among-lineage variation in diversification rates during the radiation of Australia's largest vertebrate clade. Proc. Roy. Soc. Lond. Ser. B 274:2915-2923.
}
\examples{


}
\keyword{datasets}
