\name{skinktree}
\alias{skinktree}
\docType{data}
\title{ Phylogenetic tree of Australian sphenomorphine skink genera}
\description{ 
Phylogenetic tree of 23 Australian sphenorphine skinks - including at least one representative of all genera.
  }
\usage{data(skinktree)}

\source{
Rabosky, D. L., S. C. Donnellan, A. L. Talaba, and I. J. Lovette. 2007. Exceptional among-lineage variation in diversification rates during the radiation of Australia's largest vertebrate clade. Proc. Roy. Soc. Lond. Ser. B 274:2915-2923.
}
\references{
Rabosky, D. L., S. C. Donnellan, A. L. Talaba, and I. J. Lovette. 2007. Exceptional among-lineage variation in diversification rates during the radiation of Australia's largest vertebrate clade. Proc. Roy. Soc. Lond. Ser. B 274:2915-2923.
}
\examples{


}
\keyword{datasets}