% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothie.R
\name{smooth.signal}
\alias{smooth.signal}
\title{Smooth Curve}
\usage{
smooth.signal(
  data,
  stage = "mbc",
  l = 6,
  measurements = c(),
  clear.outliers = FALSE,
  gaussian = TRUE
)
}
\description{
@param x Input data
       stage At what stage chould the data be smoothed? "X","BG","Mean","corr", tbc
       l Length of the gaussian to smooth with, corresponds to the Level of smoothness
       measurements Vector with measurement numbers that should be smoothed.
       clear.outliers Boolean to indicate whether outliers should be cleared before smoothing.
}
\examples{
clear.outliers(x = data)
clear.outliers(x = data, a = 2, use.quartile = T)
}
\keyword{outlier,}
\keyword{range}
\keyword{sigma,}
