% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothieSD.R
\name{smooth.signal.SD}
\alias{smooth.signal.SD}
\title{Smooth Curve}
\usage{
smooth.signal.SD(
  data,
  stage = "mbc",
  l = 6,
  measurements = c(),
  clear.outliers = FALSE
)
}
\description{
@param x Input data
 @param stage At what stage chould the data be smoothed? "X","BG","Mean","corr", tbc
 @param  l Length of the gaussian to smooth with, corresponds to the Level of smoothness
   @param    measurements Vector with measurement numbers that should be smoothed.
    @param   clear.outliers Boolean to indicate whether outliers should be cleared before smoothing.
}
\examples{
clear.outliers(x = data)
clear.outliers(x = data, a = 2, use.quartile = T)
}
\keyword{outlier,}
\keyword{range}
\keyword{sigma,}
