% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.background.R
\name{subtract.background}
\alias{subtract.background}
\title{Subtract background}
\usage{
subtract.background(data)
}
\arguments{
\item{data}{Data input}
}
\description{
This function subtracts the mean background off of each measured value and adds new columns with the new data.
}
\examples{
subtract.background(data)
}
\keyword{background}
