\name{warblers}
\alias{warblers}
\docType{data}
\title{ Branching times from phylogeny of North American Dendroica warblers}
\description{
  Branching times from ultrametric phylogenetic tree for Dendroica wood-warblers
}
\usage{data(warblers)}

\source{
  Lovette, I.J., E. Bermingham. 1999.
  Explosive speciation in the New World Dendroica warblers. \emph{Proc. Roy. Soc. B. Lond.} 266:1629-1636.
}
\references{
  Lovette, I.J., E. Bermingham. 1999.
  Explosive speciation in the New World Dendroica warblers. \emph{Proc. Roy. Soc. B. Lond.} 266:1629-1636.

  Rabosky, D. L., Lovette, I. J. 2008. Explosive evolutionary radiations: decreasing speciation or increasing extinction through time? \emph{Evolution} In press. 

}
\examples{
data(warblers)
plotLtt(warblers)

}
\keyword{datasets}
