% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.multiple.csv.R
\name{write.multiple.csv}
\alias{write.multiple.csv}
\title{Update this Package}
\usage{
write.multiple.csv(list, folder, filename)
}
\arguments{
\item{list}{list of data frames}

\item{folder}{name of the folder where the files should be stored}

\item{filename}{default filename}
}
\description{
This function updates the LaseR-package
}
\examples{
write.multiple.csv(list = data, folder = "new", filename = "results")
}
\keyword{update,change}
