% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xerrbar.R
\name{xerrbar}
\alias{xerrbar}
\title{Adds x error bars to plot}
\usage{
xerrbar(x, y, xsd, color, cap = 0.01, ...)
}
\arguments{
\item{x}{List of x-Values
y List of y-values
xsd List of Errors of X
color Color of the error bars
cap Length of the caps in respect to width of plot window, Default is 0.01}
}
\description{
says it all
}
\examples{
xerrbar(x,y,xsd,color="red", cap =0.01)
}
\keyword{errbar,}
\keyword{x}
