% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yerrbar.R
\name{yerrbar}
\alias{yerrbar}
\title{Adds y error bars to plot}
\usage{
yerrbar(x, y, ysd, color, cap = 0.01, ...)
}
\arguments{
\item{x}{List of x-Values}

\item{y}{List of y-values
@param       xsd List of Errors of y
 @param      color Color of the error bars
 @param      cap Length of the caps in respect to width of plot window, Default is 0.01}
}
\description{
says it all
}
\examples{
yerrbar(x,y,ysd,color="red", cap =0.01)
}
\keyword{errbar,}
\keyword{y}
