% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.ratios.R
\name{add.ratios}
\alias{add.ratios}
\title{Add Ratios}
\usage{
add.ratios(data)
}
\arguments{
\item{data}{Data frame with original data.}
}
\description{
This function adds measured and activity ratios to your data frame.
}
\examples{
add.ratios(data)
}
\keyword{ratios}

