% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.measurements.R
\name{average.measurements}
\alias{average.measurements}
\title{Average measurements}
\usage{
average.measurements(data, stage = "BG", clear.outliers = TRUE)
}
\arguments{
\item{data}{Data input}
}
\description{
This function averages the measurements and adds new columns with the new data and the standard deviation.
}
\examples{
average.measurements(data)
}
\keyword{average,}
\keyword{mean,}
\keyword{sd}
