% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_exon_couplings.R
\name{calculate_exon_couplings}
\alias{calculate_exon_couplings}
\title{Title}
\usage{
calculate_exon_couplings(junctions, reference_junctions)
}
\arguments{
\item{pathJunctionRef}{}
}
\value{

}
\description{
Title
}
