% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.outliers.R
\name{clear.outliers}
\alias{clear.outliers}
\title{Clear Outliers}
\usage{
clear.outliers(x, a = 1.5, use.quartile = T, fill.gaps = F)
}
\arguments{
\item{x}{Input data
a Sigma range
use.quartile Boolean if quartile range (TRUE,default) or standard deviation (FALSE) should be used.
fill.gaps Boolean if gaps should be filled with the mean value of neighbouring data points}
}
\description{
This function deletes the entries which deviate more than a specific sigma range from the mean.
}
\examples{
clear.outliers(x = data)
clear.outliers(x = data, a = 2, use.quartile = T)
}
\keyword{outlier,}
\keyword{range}
\keyword{sigma,}
