% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_exon_tss_couplings.R
\name{compute_exon_tss_couplings}
\alias{compute_exon_tss_couplings}
\title{Calculate couplings between exon and TSSs}
\usage{
compute_exon_tss_couplings(junctions, junctionRef)
}
\arguments{
\item{junctions:}{junctions obtained from Long read data assigned to read_ids}

\item{junctionRef:}{set of junctions of interest classified using SaiLoR}
}
\value{

}
\description{
Calculate couplings between exon and TSSs
}
