% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sailor_full_lengths.R
\name{countLinks}
\alias{countLinks}
\title{Count full length reads}
\usage{
countLinks(alignments, linksDatabase)
}
\arguments{
\item{linksDatabase}{}
}
\value{

}
\description{
Count full length reads
}
