% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference_junctions.R
\name{create_reference_junctions}
\alias{create_reference_junctions}
\title{Title}
\usage{
create_reference_junctions(pathSJ.out, min.jcounts, refAnnot, type)
}
\arguments{
\item{min.jcounts}{: minimal counts for junction filtering}

\item{refAnnot}{: reference annotation genomic ranges object}

\item{pathSJ.out:}{object of junctions from STAR}

\item{type:}{specicify type if want to use short read correction.}
}
\value{

}
\description{
Title
}
