% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference_junctions.R
\name{get_junctions_from_annot}
\alias{get_junctions_from_annot}
\title{Annotate junctions}
\usage{
get_junctions_from_annot(
  AnnotJunctions,
  refAnnot,
  linksDatabase,
  genesMultipleJunctions
)
}
\value{

}
\description{
Annotate junctions
}
