% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference_junctions.R
\name{get_junctions_from_short_reads}
\alias{get_junctions_from_short_reads}
\title{Get junctions from short read sequencing data}
\usage{
get_junctions_from_short_reads(
  pathSJ.out,
  min.counts,
  refAnnot,
  linksDatabase,
  genesMultipleJunctions
)
}
\value{

}
\description{
Get junctions from short read sequencing data
}
