% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sailor_full_lengths.R
\name{get_sailor_full_lengths}
\alias{get_sailor_full_lengths}
\title{Get sailor full lengths}
\usage{
get_sailor_full_lengths(annotPath, alignmentsFile, tss.ntwindow, tes.ntwindow)
}
\arguments{
\item{tes.ntwindow}{}
}
\value{

}
\description{
Get sailor full lengths
}
