\name{laser-internals}
\alias{laser-internals}
\alias{yuleint2}
\alias{combinations}
\alias{Ibd}
\alias{Irvbd}
\alias{IDDX}
\alias{IDDL}
\alias{IpureBirth}
\alias{Iyule2rate}
\alias{Iyule3rate}
\alias{checkbasal}



\title{ internal functions for package laser }
\description{
  internal functions for \code{laser}
}

\details{
  These functions are for internal use and should not be called by the user
}


\author{ Dan Rabosky \email{DLR32@cornell.edu} }

\keyword{ internal }
