% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaded_data_2_list.R
\name{loaded_data_2_list}
\alias{loaded_data_2_list}
\title{Average measurements}
\usage{
loaded_data_2_list(data, preablation = T)
}
\arguments{
\item{data}{Data input}

\item{preablation}{Boolean if True preablation is discarded}
}
\description{
This function averages the measurements and returns a list of average values and standard deviations.
}
\examples{
loaded_data_2_list(data = read.csv("results.csv"), preablation = T)
}
\keyword{average,}
\keyword{mean,}
\keyword{sd}
