% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference_junctions.R
\name{make_annotation_txf}
\alias{make_annotation_txf}
\title{Extract junctions from annotation}
\usage{
make_annotation_txf(refAnnot)
}
\value{

}
\description{
Extract junctions from annotation
}
