% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_to_junctions.R
\name{read_refjunction_correction}
\alias{read_refjunction_correction}
\title{Title}
\usage{
read_refjunction_correction(
  bamJunctions,
  shortReadJunctions,
  dist.5ss,
  dist.3ss
)
}
\arguments{
\item{dist.5ss}{: distance to 5ss splice site to compare to reference to be corrected in the long reads}

\item{dist.3ss}{: distance to 3ss splice site to compare to reference to be corrected in the long reads}

\item{bamJunctions:}{bam junctions obtained from read_to_junctions functions}

\item{shortReadJunctions:}{read junctions from STAR}
}
\value{

}
\description{
Title
}
