% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_to_junctions.R
\name{read_to_junctions}
\alias{read_to_junctions}
\title{junctions to reads assignments}
\usage{
read_to_junctions(bamPath, refJunAnnot, pathAnnot)
}
\arguments{
\item{bamPath:}{path to minimap2 bam file}

\item{refJunAnnot:}{reference_junctions: set of junctions of interest classified using SaiLoR}

\item{pathAnnot:}{path to reference annotation}
}
\value{

}
\description{
junctions to reads assignments
}
